requirements.txt
pandas==2.1.4
numpy==1.26.4
streamlit==1.32.2
networkx==3.2.1
plotly==5.20.0
matplotlib==3.8.4
scipy==1.12.0
requests==2.31.0
python-dotenv==1.0.1